Using a super-simple approach for scrolling that doesn't require the complex nested scrolling behavior by leveraging UIScrollView in Obj-C.
This fixes a number of issues: Scroll-To-Top on status-bar tap works properly (YES!), the page does not scroll any more at all, only the content-container does because the scrolling is disabled for the main UIScrollView and it also always has the proper bouncing behavior on top/bottom. It works now *exactly* like it does in a native app.